modified server.js

+++ server.js
@@ -12,13 +12,20 @@ app.get('/api/getLP/:imgID', (req, res) => {
     exec(`wget https://i.imgur.com/${image}.jpg`);
 
     exec(`alpr ${image}.jpg`, (err, stdout, stderr) => {
+try{
         var output = stdout.split('\n')[1].match('[A-Z0-9]{3,6}')[0];
+} catch (e) {
+	exec(`rm ${image}.jpg`);
+	res.json({"license_plate":"NO PLATE DETECTED"});
+};
         
-        exec(`rm ${image}.jpg`);
+if(output){
+	exec(`rm ${image}.jpg`);
         res.json({"license_plate":`${output}`});
+};
     });
 
-    res.json(200);
+    //res.json(200);
 });
 
-app.listen(port, () => console.log("App started..."));
+app.listen(port,() => console.log("App started..."));
